+++ shuup_tests/admin/test_contact_page.py
@@ -22,17 +22,17 @@ from shuup.testing.utils import apply_request_middleware
 
 @pytest.mark.django_db
 def test_contact_details_view_with_many_groups(rf, admin_user):
-    get_default_shop()
+    shop = get_default_shop()
     person = create_random_person()
     person.groups.add(
-        ContactGroup.objects.create(name="Czz Group"),
-        ContactGroup.objects.create(name="Azz Group"),
-        ContactGroup.objects.create(name="Bzz Group"),
-        ContactGroup.objects.language('fi').create(name="Dzz ryhmä"),
+        ContactGroup.objects.create(name="Czz Group", shop=shop),
+        ContactGroup.objects.create(name="Azz Group", shop=shop),
+        ContactGroup.objects.create(name="Bzz Group", shop=shop),
+        ContactGroup.objects.language('fi').create(name="Dzz ryhmä", shop=shop),
     )
 
     # Group with name in two languages
-    grp_e = ContactGroup.objects.language('en').create(name="Ezz Group")
+    grp_e = ContactGroup.objects.language('en').create(name="Ezz Group", shop=shop)
     grp_e.set_current_language('fi')
     grp_e.name = "Ezz ryhmä"
     grp_e.save()